+++ src/rard/research/tests/models/test_original_text.py
@@ -3,15 +3,12 @@ from django.db.utils import IntegrityError
 from django.test import TestCase
 
 from rard.research.models import (
-    Antiquarian,
     CitingWork,
-    Concordance,
     Fragment,
     OriginalText,
     Reference,
     Translation,
 )
-from rard.research.models.base import FragmentLink
 
 pytestmark = pytest.mark.django_db
 
@@ -51,52 +48,6 @@ class TestOriginalText(TestCase):
         }
         text = OriginalText.objects.create(**data, owner=self.fragment)
         self.assertEqual(text.remove_reference_order_padding(), "1.10.12345")
-
-
-class TestConcordance(TestCase):
-    def setUp(self):
-        self.fragment = Fragment.objects.create(name="name")
-        self.citing_work = CitingWork.objects.create(title="title")
-        self.original_text = OriginalText.objects.create(
-            owner=self.fragment, citing_work=self.citing_work
-        )
-
-    def test_creation(self):
-        data = {
-            "source": "source",
-            "identifier": "identifier",
-        }
-        concordance = Concordance.objects.create(
-            **data, original_text=self.original_text
-        )
-        for key, val in data.items():
-            self.assertEqual(getattr(concordance, key), val)
-
-    def test_required_fields(self):
-        # required fields on forms
-        self.assertFalse(Concordance._meta.get_field("source").blank)
-        self.assertFalse(Concordance._meta.get_field("identifier").blank)
-
-    def test_original_text_required(self):
-        data = {
-            "source": "source",
-            "identifier": "identifier",
-        }
-        # cannot create a concordance with no original text
-        with self.assertRaises(IntegrityError):
-            Concordance.objects.create(**data)
-
-    def test_concordance_identifiers(self):
-        # test the names to go in the concordance table including ordinals
-        # set up a second original text
-        OriginalText.objects.create(owner=self.fragment, citing_work=self.citing_work)
-        a = Antiquarian.objects.create(name="name1", re_code="name1")
-        link = FragmentLink.objects.create(fragment=self.fragment, antiquarian=a)
-
-        concordances = link.get_concordance_identifiers()
-        self.assertEqual(
-            [x[0] for x in self.original_text.concordances.all()], concordances
-        )
 
 
 class TestTranslation(TestCase):
